# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HAMSAM",
        "description": "\n * \t  1) call login method and get access_token add token_type in this format : {token_type}+{access_token}\n * \t  2) use them as authorization keys",
        "contact": {
            "name": "alizand",
            "email": "alizandd@gmail.com"
        },
        "version": "1"
    },
    "servers": [
        {
            "url": "{Schema}://{Host}/{Version}",
            "description": "DEVELOP",
            "variables": {
                "Schema": {
                    "enum": [
                        "https",
                        "http"
                    ],
                    "default": "http"
                },
                "Version": {
                    "enum": [
                        "v1",
                        "v2"
                    ],
                    "default": "v1"
                },
                "Host": {
                    "enum": [
                        "localhost/backend/src/public/api",
                        "ssn.tvapps.ir/api",
                        "127.0.0.1:8000/api"
                    ],
                    "default": "localhost/backend/src/public/api"
                }
            }
        }
    ],
    "paths": {
...
        "/wall/posts/{postId}/comments": {
            "post": {
                "tags": [
                    "Wall"
                ],
                "summary": "Submit a comment on a post with optional media",
                "operationId": "storeCommentWithMedia",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "ID of the post to comment on",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Submit a comment with or without media files.",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "content": {
                                        "description": "Comment content",
                                        "type": "string"
                                    },
                                    "media[]": {
                                        "description": "Media files",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Comment created successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Post not found"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
        "/wall/likes/toggle": {
            "post": {
                "tags": [
                    "Wall"
                ],
                "summary": "Toggle like on a post or comment",
                "operationId": "toggleLike",
                "requestBody": {
                    "description": "Toggle like on an item",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "id"
                                ],
                                "properties": {
                                    "type": {
                                        "description": "Type of the item to like",
                                        "type": "string",
                                        "enum": [
                                            "Post",
                                            "Comment"
                                        ],
                                        "example": "Post"
                                    },
                                    "id": {
                                        "description": "ID of the item to like",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Like toggled successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully liked"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized ."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
        "/wall/posts": {
            "get": {
                "tags": [
                    "Wall"
                ],
                "summary": "Show All posts",
                "operationId": "indexPost",
                "responses": {
                    "200": {
                        "description": "Post created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "اطلاعات شما با موفقیت ثبت شد ."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or expired token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "کد وارد شده نامعتبر یا منقضی شده است ."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Wall"
                ],
                "summary": "Create a new post",
                "operationId": "createPost",
                "requestBody": {
                    "description": "Create a new post with content and optional files",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "content"
                                ],
                                "properties": {
                                    "content": {
                                        "description": "The content of the post",
                                        "type": "string",
                                        "example": "This is a sample post content."
                                    },
                                    "media": {
                                        "description": "Optional array of files (images/videos)",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Post created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "اطلاعات شما با موفقیت ثبت شد ."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or expired token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "کد وارد شده نامعتبر یا منقضی شده است ."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Authorization": {
                "type": "apiKey",
                "description": "{token_type}+{access_token}",
                "name": "Authorization",
                "in": "header",
                "default": "Bearer {apiKey}"
            }
        }
    }
}